fix: 로딩경로 수정및 삭제 #162
