+++ src/app/(root)/eventsdetail/loading.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import dynamic from "next/dynamic";
 import animationData from "@/assets/loadingBar.json";
 
-const LottiAnimation = dynamic(() => import("@/components/Loading/LottiAnimation"), {
+const LottiAnimation = dynamic(() => import("@/components/Common/Loading/LottiAnimation"), {
   ssr: false,
 });
