+++ src/app/(root)/maindetail/[id]/page.tsx
@@ -15,7 +15,7 @@ import ShareButton from "@/components/MainDetail/ShareButton";
 import dynamic from "next/dynamic";
 import animationData from "@/assets/loadingBar.json";
 
-const LottiAnimation = dynamic(() => import("@/components/Loading/LottiAnimation"), {
+const LottiAnimation = dynamic(() => import("@/components/Common/Loading/LottiAnimation"), {
   ssr: false,
 });
