+++ src/components/Common/Loading/LottiAnimation.tsx
@@ -10,18 +10,15 @@ interface LottiAnimationProps {
   className?: string;
 }
 
-const LottiAnimation: React.FC<LottiAnimationProps> = ({
-  animationData,
-  size = "100%",
-  isFixed = false,
-  className = "",
-}) => {
+const LottiAnimation: React.FC<LottiAnimationProps> = ({ animationData, size = '100%', isFixed = false, className = '' }) => {
   return (
-    <div
-      className={`${isFixed ? "fixed inset-0" : "relative"} flex items-center justify-center ${className}`}
-      style={{ width: size, height: size }}
-    >
-      <Lottie loop animationData={animationData} play style={{ width: "100%", height: "100%" }} />
+    <div className={`${isFixed ? 'fixed inset-0' : 'relative'} flex items-center justify-center ${className}`} style={{ width: size, height: size }}>
+      <Lottie
+        loop
+        animationData={animationData}
+        play
+        style={{ width: '100%', height: '100%' }}
+      />
     </div>
   );
 };
